requirements.txt
Flask==3.0.0
boto3==1.34.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.7
gunicorn==21.2.0
